feat: add usage of db to getProductById lambda
